requirements.txt
jishaku
toml~=0.10.1
uvloop~=0.15.0
parsedatetime~=2.6
pytz~=2021.1
discord.py-stubs~=1.6
#discord.py[voice]>=1.5.0
git+https://github.com/Rapptz/discord.py.git
tortoise-orm[accel]>=0.16.21
aerich
asyncpg~=0.22.0
python-rapidjson
ciso8601
Babel==2.9.0
#git+https://github.com/paris-ci/babel.git
git+https://github.com/Rapptz/discord-ext-menus
python-dateutil~=2.8.1
aiohttp[speedups]~=3.7.3
aiohttp_cors~=0.7.0
polib~=1.1.0
tqdm~=4.56.2
pillow>=8.0
statcord.py~=3.0.8